Store and reuse transcriptions, changed default model to LLaMa-3

+++ summarize.sh
@@ -8,7 +8,8 @@
 
 
 OLLAMA_HOST=127.0.0.1:11434
-OLLAMA_MODEL=mixtral
+OLLAMA_MODEL=llama3
+#OLLAMA_MODEL=mixtral
 OLLAMA_CONTEXT_SIZE=8192
 # >80GB RAM needed to use the full context width of Mixtral
 #OLLAMA_CONTEXT_SIZE=32768
@@ -17,7 +18,8 @@ WHISPER_CPP_MODEL=ggml-large-v3.bin
 WHISPER_CPP_NTHREADS=4
 WHISPER_CPP_NPROCESSORS=1
 # Possible speedup to the transcribing phase
-#WHISPER_CPP_NPROCESSORS=4
+#WHISPER_CPP_NPROCESSORS=3
+STORAGE_DIR=".summarize.data"
 
 function help {
 	echo "Usage: summarize.sh [-q] [-c] [-d] <Youtube url>"
@@ -26,6 +28,7 @@ function help {
 	echo
 	echo "  -q : Quiet, does not print any progress information or execution time"
 	echo "  -c : Conversational, waits for further questions after the summary is printed"
+	echo "  -C : Conversational, immediately prompts without first doing a summary"
 	echo "  -d : Debug, prints all output from the called programs"
 	echo
 }
@@ -47,8 +50,12 @@ function unload {
 }
 
 function cleanup {
-	rm ${orig_audio}.wav
-	rm ${audio}.wav
+	if [ ! -z ${orig_audio} ]; then
+		rm ${orig_audio}.wav
+	fi
+	if [ ! -z ${audio} ]; then
+		rm ${audio}.wav
+	fi
 }
 
 function print {
@@ -65,7 +72,7 @@ function error {
 function query {
 	arg="{";
 	arg+="\"model\": \"$OLLAMA_MODEL\","
-	arg+="\"prompt\": \"$1\","
+	arg+="\"prompt\": \"$@\","
 	arg+="\"stream\": false,"
 	if [ ! -z "$context" ]; then
 		arg+="\"context\": ${context},"
@@ -93,6 +100,10 @@ function query {
 	if [ true = $fail ]; then
 		cleanup
 		exit
+	fi
+	notokens=$(echo ${resp} | jq '.prompt_eval_count')
+	if (( $notokens > $OLLAMA_CONTEXT_SIZE )); then
+		resp+="WARNING: The size of the input to the LLM exceeded its configured max size! Output might be unreliable.\n\n"
 	fi
 	echo "$resp"
 }
@@ -123,6 +134,10 @@ while (( "$#" )); do
 			;;
 		-c )
 			conv=true
+			;;
+		-C )
+			conv=true
+			skipsum=true
 			;;
 		-d )
 			debug=true
@@ -174,25 +189,41 @@ fi
 
 start=$(date +%s)
 
-orig_audio=$(mktemp -u)
-print "Downloading audio... "
-if [ true = $debug ]; then
-	yt-dlp -x --audio-format wav -o ${orig_audio}.wav "$video"
+# create storage directory if it doesn't exist
+mkdir -p "$STORAGE_DIR"
+# create a hash from the URL to store/recall data
+storage=$(echo $video | md5sum - | awk '{print $1}')
+if [ -d "$STORAGE_DIR/$storage" ]; then
+	print "Existing data found for this URL ($storage) - reusing"
 else
-	yt-dlp -x --audio-format wav -o ${orig_audio}.wav "$video" &>/dev/null
+	mkdir -p "$STORAGE_DIR/$storage"
+	echo "$video" > "$STORAGE_DIR/$storage/url.txt"
 fi
-print "Converting audio... "
-audio=$(mktemp -u)
-if [ true = $debug ]; then
-	ffmpeg -i ${orig_audio}.wav -ar 16000 -ac 1 -c:a pcm_s16le ${audio}.wav
+if [ -f "$STORAGE_DIR/$storage/transcription.txt" ]; then
+	print " transcribed audio"
+	input=$(cat "$STORAGE_DIR/$storage/transcription.txt")
 else
-	ffmpeg -i ${orig_audio}.wav -ar 16000 -ac 1 -c:a pcm_s16le ${audio}.wav &>/dev/null
-fi
-print "Transcribing audio... "
-if [ true = $debug ]; then
-	input=$($WHISPER_CPP_PATH/main -p $WHISPER_CPP_NPROCESSORS -t $WHISPER_CPP_NTHREADS -m $WHISPER_CPP_PATH/models/$WHISPER_CPP_MODEL -np -nt -f ${audio}.wav)
-else
-	input=$($WHISPER_CPP_PATH/main -p $WHISPER_CPP_NPROCESSORS -t $WHISPER_CPP_NTHREADS -m $WHISPER_CPP_PATH/models/$WHISPER_CPP_MODEL -np -nt -f ${audio}.wav 2>/dev/null)
+	orig_audio=$(mktemp -u)
+	print "Downloading audio... "
+	if [ true = $debug ]; then
+		yt-dlp -x --audio-format wav -o ${orig_audio}.wav "$video"
+	else
+		yt-dlp -x --audio-format wav -o ${orig_audio}.wav "$video" &>/dev/null
+	fi
+	print "Converting audio... "
+	audio=$(mktemp -u)
+	if [ true = $debug ]; then
+		ffmpeg -i ${orig_audio}.wav -ar 16000 -ac 1 -c:a pcm_s16le ${audio}.wav
+	else
+		ffmpeg -i ${orig_audio}.wav -ar 16000 -ac 1 -c:a pcm_s16le ${audio}.wav &>/dev/null
+	fi
+	print "Transcribing audio... "
+	if [ true = $debug ]; then
+		input=$($WHISPER_CPP_PATH/main -p $WHISPER_CPP_NPROCESSORS -t $WHISPER_CPP_NTHREADS -m $WHISPER_CPP_PATH/models/$WHISPER_CPP_MODEL -np -nt -f ${audio}.wav)
+	else
+		input=$($WHISPER_CPP_PATH/main -p $WHISPER_CPP_NPROCESSORS -t $WHISPER_CPP_NTHREADS -m $WHISPER_CPP_PATH/models/$WHISPER_CPP_MODEL -np -nt -f ${audio}.wav 2>/dev/null)
+	fi
+	echo "$input" > $STORAGE_DIR/$storage/transcription.txt
 fi
 if [ -z "${input}" ]; then
 	error
@@ -201,19 +232,24 @@ if [ -z "${input}" ]; then
 	cleanup
 	exit
 fi
-print "Summarizing text... "
-input=${input//\"/\\\"}
-prompt="Please summarize the following textual contents of a video: $input"
-resp=$(query "$prompt")
-summary=$(echo ${resp} | jq '.response')
-context=$(echo ${resp} | jq '.context')
+if [ -z $skipsum ]; then
+	print "Summarizing text... "
+	input=${input//\"/\\\"}
+	prompt="Please summarize the following audio transcription: $input"
+	resp=$(query "$prompt")
+	summary=$(echo ${resp} | jq '.response')
+	context=$(echo ${resp} | jq '.context')
+fi
 
 end=$(date +%s)
-print "Completed in $(($end-$start)) seconds."
 
-echo
-echo
-echo -e $summary
+if [ -z $skipsum ]; then
+	print "Completed in $(($end-$start)) seconds."
+	echo
+	echo
+	echo -e $summary
+fi
+
 echo
 cleanup
 
@@ -230,7 +266,7 @@ while true; do
 	if [ -z "$question" ]; then
 		break
 	else
-		resp=$(query "$question. Make sure to only use knowledge found in the following textual representation of a video: $input")
+		resp=$(query "$question. Make sure to only use knowledge found in the following audio transcription: $input")
 		answer=$(echo ${resp} | jq '.response')
 		context=$(echo ${resp} | jq '.context')
 		echo -e $answer
